Features: all test cases passed

+++ Calculator/Multipule.py
@@ -1,2 +1,2 @@
 def mul(a, b):
-    return int(a) * int(b)
+    return float(a) * float(b)
