+++ Calculator/Square.py
@@ -1,4 +1,3 @@
 def sq(a):
-    # I think we should remove int cast
-    return int(a) ** 2
+    return float(a) ** 2
